requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.92.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
duckdb==1.5.1
exceptiongroup==1.3.1
fastapi==0.135.3
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
numpy==2.2.6
openai==2.31.0
packaging==26.0
pandas==2.3.3
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pymongo==4.16.0
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
ruff==0.15.10
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tomli==2.4.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
